requirements.txt
Flask
numpy
pandas
pickle5
Pillow
streamlit
scikit_learn